Secure forms abstraction final draft- doy, security already loaded and avoid using $_POST- doy!

+++ system/expressionengine/libraries/EE_Security.php
@@ -43,16 +43,41 @@ class EE_Security extends CI_Security {
 	 *
 	 * @access	public
 	 * @param 	string
-	 * @param 	string
 	 * @return	bool
 	 */
 	function secure_forms_check($xid)
 	{	
+		if ( ! $this->check_xid($xid))
+		{
+			return FALSE;
+		}
+		
+  		$this->delete_xid($xid);
+
+		return TRUE;
+	}
+	
+	// --------------------------------------------------------------------
+
+	/**
+	 * Check for Valid Security Hash
+	 *
+	 * @access	public
+	 * @param 	string
+	 * @return	bool
+	 */
+	function check_xid($xid)
+	{
 		if ($this->EE->config->item('secure_forms') != 'y')
 		{
 			return TRUE;
 		}
 		
+		if ( ! $xid)
+		{
+			return FALSE;
+		}		
+
 		$this->EE->db->where('hash', $xid);
 		$this->EE->db->where('ip_address', $this->EE->input->ip_address());
 		$this->EE->db->where('date > UNIX_TIMESTAMP()-7200');
@@ -63,13 +88,34 @@ class EE_Security extends CI_Security {
 		{
 			return FALSE;
 		}
+		
+		return TRUE;		
+	}
+	
+	// --------------------------------------------------------------------
+
+	/**
+	 * Delete Security Hash
+	 *
+	 * @access	public
+	 * @param 	string
+	 * @return	void
+	 */
+	
+	function delete_xid($xid)
+	{
+		if ($this->EE->config->item('secure_forms') != 'y' OR $xid == FALSE)
+		{
+			return;
+		}
 
 		$this->EE->db->where("(hash='".$this->EE->db->escape_str($xid)."' AND ip_address = '".$this->EE->input->ip_address()."')", NULL, FALSE);
 		$this->EE->db->or_where('date < UNIX_TIMESTAMP()-7200');
 		$this->EE->db->delete('security_hashes');
 		
-		return TRUE;
+		return;		
 	}
+
 }
 // END CLASS
